+++ pages/api/user/[userId].ts
@@ -2,11 +2,21 @@ import { User } from '@prisma/client';
 import { NextApiRequest, NextApiResponse } from 'next';
 import { errorResponse, successResponse } from '@/lib/http.response';
 import { prisma } from '@/lib/prisma';
+import apiAuth from '@/lib/apiAuth';
 
 export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse
 ) {
+  const isAuth = await apiAuth(req);
+  if (!isAuth) {
+    return errorResponse({
+      res,
+      message: 'Unauthorized',
+      statusCode: 401,
+      success: false,
+    });
+  }
   switch (req.method) {
     case 'GET':
       const { userId } = req.query;
